[9] - Refactorizacion de ficheros redundantes y naming
