+++ src/App.test.tsx
@@ -1,8 +1,10 @@
 import { render, screen } from '@testing-library/react'
 import App from './App'
 
-it('muestra un gif en pantalla', async () => {
-  render(<App />)
+describe('listado de gifs', () => {
+  it('muestra un gif en pantalla con texto alternativo', async () => {
+    render(<App />)
 
-  expect(await screen.findByAltText('perro loco')).toBeVisible()
+    expect(await screen.findByAltText('perro loco')).toBeVisible()
+  })
 })
