--- src/styles/index.js
@@ -1,3 +1,0 @@
-export * from "./reset.css";
-export * from "./reboot.css";
-export * from "../components/Gif.css"
