+++ src/types/interface.ts
@@ -1,4 +1,4 @@
-export interface gifType {
+export interface Gif {
   name: string
   imgUrl: string
 }
